♻️ Only disable confirmation logic for onboarding
