+++ modules/ppcp-settings/resources/js/Components/Screens/Onboarding/Onboarding.js
@@ -3,12 +3,27 @@ import { OnboardingHooks } from '../../../data';
 
 import { getSteps, getCurrentStep } from './availableSteps';
 import Navigation from './Components/Navigation';
+import { useEffect } from '@wordpress/element';
 
 const Onboarding = () => {
 	const { step, setStep, setCompleted, flags } = OnboardingHooks.useSteps();
 
 	const Steps = getSteps( flags );
 	const currentStep = getCurrentStep( step, Steps );
+
+	// Disable the "Changes you made might not be saved" browser warning.
+	useEffect( () => {
+		const suppressBeforeUnload = ( event ) => {
+			event.stopImmediatePropagation();
+			return undefined;
+		};
+
+		window.addEventListener( 'beforeunload', suppressBeforeUnload );
+
+		return () => {
+			window.removeEventListener( 'beforeunload', suppressBeforeUnload );
+		};
+	}, [] );
 
 	const handleNext = () => setStep( currentStep.nextStep );
 	const handlePrev = () => setStep( currentStep.prevStep );
